Set commit info using create or update
Due to how the read func was written Terraform will try to find the
commit info each time it does a refresh. As you cannot get this info
easily from the GitHub API, the provider would loop through commits
until it found one containing the file in question. When the commit is a
a distant memory, it can incur many API requests and, when managing many
files across many repositories, could result in being rate limited.

To address this instead we now set this info when the file is created or
updated and not during the refresh. This is pretty safe as for this info
to change upstream it would mean changing the file contents, which would
trigger Terraform to update the file again (to enforce desired state).

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>