Correct documentation of provider configuration, and deprecate organization argument (#735)
* Remove dead function

* Remove redundant call to os.Getenv

* Simplify github.Config struct

- Remove 'Organization' field, which was never read
- Remove 'Individual' field, which was never read
- Replace 'Anonymous' field with a function (as the field is derived
  from the value of 'Token')

* Deprecate organization and GITHUB_ORGANIZATION

We don't need `organization` and `GITHUB_ORGANIZTION`, because the
current behaviour is that the user can put a user name _or_ an
organization name in any of `owner`, `organization`, `GITHUB_OWNER`, or
`GITHUB_ORGANIZATION`.

It therefore seems clearer to just keep `owner` and `GITHUB_OWNER`.

Unfortunately, the current precedence of owner, GITHUB_OWNER,
organization, GITHUB_ORGANIZATION is also a bit confusing. From high
precedence to low precedence, the order is:

- organization
- GITHUB_ORGANIZATION
- GITHUB_OWNER
- owner

For backwards compatibility, in minor/patch releases we need to preserve
this order. However, I've contained the code that does this in a single
clearly marked block of code in providerConfigure so that it can be
removed easily in a future major release.

I've updated the documentation with:

- a description of the current behaviour
- the fact that organization and GITHUB_ORGANIZATION are deprecated
- a warning that the precedence of owners/GITHUB_OWNER may change