fix: use the appropriate ID when trying to import `github_team_members` objects (#1074)
When importing a `github_team_member`, the import command passes in team ID as the
"Id", not the "team_id" captured in the state for existing resources, and this empty string is
causing the import step to fail when parsing as a number. Should fix #608, as I hit the same error.