Use correct base url for `gh auth token --hostname`
This PR is a follow-up to https://github.com/integrations/terraform-provider-github/pull/1854

GitHub CLI uses different base URLs in `~/.config/gh/hosts.yml`, so when we're using the standard base path of this provider, it doesn't align with how `gh` CLI stores the credentials. The following doesn't work:

```
$ gh auth token --hostname api.github.com
> no oauth token
```
... but the following does work correctly

```
$ gh auth token --hostname github.com
> gh..<valid token>
```

attempted a workaround with

```
provider "github" {
  owner = "..."
  base_url = "github.com"
}
```

But it didn't work as expected.