Fixed invalid default hostname for gh auth token
Currently in the event that no base_url is given to the provider configuration, the tokenFromGhCli() function will set a default hostname of "api.github.com" which is invalid for the "gh auth token" cli command. Instead, the proper default hostname should be "github.com" which is correct and works.